refactor: use Linux kernel time_after pattern for scroll wrap handling
time_after(a,b) = (a-b) as i32 > 0 — compares difference, not absolute.
Works correctly even when u32 wraps, as long as the interval between
compared values is less than u32::MAX/2 (~34 years at 2 ticks/sec).

+++ src/frontend.rs
@@ -35,7 +35,7 @@ enum CompletionCtx {
 
 pub struct Frontend {
     pub input: TextArea<'static>,
-    pub(crate) scroll: Cell<u64>,
+    pub(crate) scroll: Cell<u32>,
     pub(crate) follow_tail: Cell<bool>,
     pub(crate) menu_idx: usize,
     pub(crate) tab_cycle: Option<TabCycle>,
@@ -64,7 +64,7 @@ impl Frontend {
         let view = view_rx.borrow().clone();
         Self {
             input,
-            scroll: Cell::new(0u64),
+            scroll: Cell::new(0u32),
             follow_tail: Cell::new(true),
             menu_idx: 0,
             tab_cycle: None,
